Show factorial level in too long text * include factorial level for too long messages with only one factorial * moved round to math * adjusted decimals * shortened footer message * removed tolik518 from footer
