+++ src/math.rs
@@ -25,6 +25,44 @@ fn multifactorial_recursive(n: u64, k: u64, low_i: u64, high_i: u64) -> BigInt {
     }
 }
 
+
+/// Rounds a base 10 number string.
+/// Uses the last digit to decide the rounding direction.
+/// Rounds over 9s. This does **not** keep the length or turn rounded over digits into zeros.
+/// If the input is all 9s, this will round to 10.
+///
+/// # Panic
+/// This function may panic if less than two digits are supplied, or if it contains a non-digit of base 10.
+pub(crate) fn round(number: &mut String) {
+    // Check additional digit if we need to round
+    if let Some(digit) = number
+        .pop()
+        .map(|n| n.to_digit(10).expect("Not a base 10 number"))
+    {
+        if digit >= 5 {
+            let mut last_digit = number
+                .pop()
+                .and_then(|n| n.to_digit(10))
+                .expect("Not a base 10 number");
+            // Carry over at 9s
+            while last_digit == 9 {
+                let Some(digit) = number
+                    .pop()
+                    .map(|n| n.to_digit(10).expect("Not a base 10 number"))
+                else {
+                    // If we reached the end we get 10
+                    *number = "10".to_string();
+                    return;
+                };
+                last_digit = digit;
+            }
+            // Round up
+            number.push_str(&format!("{}", last_digit + 1));
+        }
+    }
+}
+
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -172,5 +210,26 @@ mod tests {
         let num = 100_001;
         let result = factorial(num, 1);
         assert_eq!(result.to_string().len(), 456579);
+    }
+
+    #[test]
+    fn test_round_down() {
+        let mut number = String::from("1929472373");
+        round(&mut number);
+        assert_eq!(number, "192947237");
+    }
+
+    #[test]
+    fn test_round_up() {
+        let mut number = String::from("74836748625");
+        round(&mut number);
+        assert_eq!(number, "7483674863");
+    }
+
+    #[test]
+    fn test_round_carry() {
+        let mut number = String::from("24999999995");
+        round(&mut number);
+        assert_eq!(number, "25");
     }
 }
